python-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# NLP & AI
spacy==3.7.2
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
nltk==3.8.1

# OpenAI (optional for GPT-based extraction)
# openai==1.3.7

# Additional utilities
requests==2.31.0
httpx==0.25.2